added authorization, main page, new event and styles

+++ projects/backend/src/utils/deleteFile.js
@@ -1,12 +1,13 @@
 const fs = require('fs');
+const { usersFolder } = require('../config');
 
 module.exports = (folder, name, s3) => {
   if (!name) {
-    console.log(`no image to delete`);
+    console.log(`no file to delete`);
     return;
   }
 
-  if (process.env.NODE_ENV === 'production') {
+  if (process.env.NODE_ENV === 'production' && folder !== usersFolder) {
     s3.deleteObject(
       {
         Bucket: process.env.AWS_BUCKET_NAME,
